refactor: Oauth2UserInfo 변경
- attributes 멤버변수를 반환해주는 getAttributes 메서드 추가
- Oauth2UserInfo를 포함한 관련된 파일 위치를 info 폴더 안으로 변경
